+++ src/ir/mod.rs
@@ -6,6 +6,7 @@ pub mod cone_function_emit;
 pub mod dedup;
 pub mod function_emit;
 pub mod generate_loop;
+pub mod multi_output_task_emit;
 pub mod param;
 pub mod soft_union;
 pub mod task_emit;